Rewrite play_card use case.
