+++ tests/test_game_service.py
@@ -196,7 +196,7 @@ class PlayerContextTest(unittest.TestCase):
 
         # when player-1 discard countess
         PlayCard().execute(
-            PlayCard.input(self.game_id, "1", "伯爵夫人", None), PlayCard.output()
+            PlayCard.input(self.game_id, "1", "伯爵夫人", None), PlayCard.presenter()
         )
 
         expected_card_mapping = {
@@ -355,7 +355,7 @@ class PlayerContextTest(unittest.TestCase):
         # when: 1對3打出神父
         PlayCard().execute(
             PlayCard.input(self.game_id, "1", "神父", ToSomeoneCard(chosen_player=3)),
-            PlayCard.output(),
+            PlayCard.presenter(),
         )
 
         status_of_player1: GameStatus = GameStatus.parse_obj(
